crawler/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dask==2023.6.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distributed==2023.6.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl#sha256=b492fd2dfeee931dd60232fd1a122bb5c3542ebcd5d614f11844647050aa539e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gensim==4.3.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
grpcio==1.54.2
h5py==3.9.0
identify==2.5.24
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.7.0
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lazy-object-proxy==1.9.0
libclang==16.0.0
llvmlite==0.40.1
locket==1.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
memory-profiler==0.61.0
mistune==3.0.1
ml-dtypes==0.2.0
msgpack==1.0.5
multidict==6.0.4
multipledispatch==0.6.0
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parse==1.19.1
parso==0.8.3
partd==1.4.0
pathy==0.10.2
peewee==3.16.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.7.0
pre-commit==3.3.3
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyee==8.2.2
Pygments==2.15.1
pylint==2.17.4
PyMySQL==1.0.3
pyparsing==3.1.0
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
tblib==2.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
tldextract==3.4.4
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
unicode==2.9
Unidecode==1.3.6
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.16
virtualenv==20.23.1
w3lib==2.1.1
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
websockets==10.4
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wordsegment==1.3.1
wrapt==1.14.1
xgboost==1.7.6
yarl==1.9.2
zict==3.0.0
zipp==3.15.0
